added poor mans debugging to generated code removed unused usings
