+++ cs/Serializer/VowpalWabbitSerializerFactory.cs
@@ -18,6 +18,8 @@ using VW.Serializer.Intermediate;
 using VW.Serializer.Reflection;
 using VW.Serializer.Visitors;
 using VW.Interfaces;
+using System.IO;
+using System.Runtime.CompilerServices;
 
 namespace VW.Serializer
 {
@@ -82,6 +84,14 @@ namespace VW.Serializer
             return newSerializer;
         }
 
+        public static Expression Log([CallerFilePath] string filePath = "", [CallerLineNumber] int lineNumber = 0)
+        {
+            var file = Expression.Constant(@"c:\vowpal_wabbit\test.log");
+            var logMethod = (MethodInfo)ReflectionHelper.GetInfo((string a) => File.AppendAllText(a,a));
+
+            return Expression.Call(logMethod, file, Expression.Constant(filePath + ":" + lineNumber + "\n"));
+        }
+
         private static Func<TExample, TVisitor, TExampleResult> CreateSerializer<TExample, TVisitor, TExampleResult>(ModuleBuilder moduleBuilder)
             where TVisitor : IVowpalWabbitVisitor<TExampleResult>
         {
@@ -102,15 +112,22 @@ namespace VW.Serializer
 
             var body = new List<Expression>();
 
+            body.Add(Log());
+
             // CODE if (value == null) throw new ArgumentNullException("value");
             body.Add(Expression.IfThen(
                     Expression.Equal(valueParameter, Expression.Constant(null)), 
                     Expression.Throw(Expression.New(ArgumentNullExceptionConstructorInfo, Expression.Constant("value")))));
 
+            body.Add(Log());
+
             // CODE if (value == null) throw new ArgumentNullException("visitor");
             body.Add(Expression.IfThen(
                     Expression.Equal(visitorParameter, Expression.Constant(null)),
                     Expression.Throw(Expression.New(ArgumentNullExceptionConstructorInfo, Expression.Constant("visitor")))));
+
+            body.Add(Log());
+
 
             var variables = new List<ParameterExpression>();
             var namespaceVariables = new List<ParameterExpression>();
@@ -155,9 +172,11 @@ namespace VW.Serializer
                     namespaceVariables.Add(namespaceVariable);
 
                     // CODE namespace = new Namespace<float> { ... };
+                    body.Add(Log());
                     body.Add(Expression.Assign(namespaceVariable, namespaceDense));
 
                     // CODE namespace.Visit = () => visitor.Visit(namespace)
+                    body.Add(Log());
                     body.Add(Expression.Assign(
                             Expression.Property(namespaceVariable, namespaceType.GetProperty("Visit")),
                             Expression.Lambda<Action>(
@@ -179,6 +198,7 @@ namespace VW.Serializer
                         featureVariables.Add(featureVariable);
 
                         // CODE feature = new Feature<float> { ... };
+                        body.Add(Log());
                         body.Add(Expression.Assign(featureVariable, feature.NewFeatureExpression));
                     }
 
@@ -195,6 +215,7 @@ namespace VW.Serializer
                     namespaceVariables.Add(namespaceVariable);
 
                     // CODE namespace = new NamespaceSparse { ... }
+                    body.Add(Log());
                     body.Add(Expression.Assign(namespaceVariable, namespaceSparse));
 
                     // loop unrolling to have dispatch onto the correct Visit<T>
@@ -219,6 +240,7 @@ namespace VW.Serializer
                         }
 
                         // CODE feature.Visit = visitor.Visit;
+                        body.Add(Log());
                         body.Add(
                             Expression.Assign(
                                 Expression.Property(featureVariable, featureVariable.Type.GetProperty("Visit")),
@@ -226,6 +248,7 @@ namespace VW.Serializer
                     }
 
                     // CODE namespace.Visit = () => { visitor.Visit(namespace); });
+                    body.Add(Log());
                     body.Add(
                         Expression.Assign(
                             Expression.Property(
@@ -260,6 +283,7 @@ namespace VW.Serializer
             }
 
             // CODE return visitor.Visit(label, new[] { ns1, ns2, ... })
+            body.Add(Log());
             body.Add(
                 Expression.Call(
                     visitorParameter,
