+++ cs_unittest/BaseData.cs
@@ -1,10 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace cs_unittest
+﻿namespace cs_unittest
 {
     public class BaseData
     {
