+++ cs_unittest/Test1and2.cs
@@ -1,19 +1,14 @@
-﻿using cs_unittest;
+﻿using System;
+using System.Collections.Generic;
+using System.Globalization;
+using System.IO;
+using System.Linq;
+using cs_unittest;
+using Microsoft.VisualStudio.TestTools.UnitTesting;
 using VW;
 using VW.Interfaces;
 using VW.Labels;
 using VW.Serializer.Attributes;
-using Microsoft.VisualStudio.TestTools.UnitTesting;
-using System;
-using System.Collections.Generic;
-using System.Collections.ObjectModel;
-using System.Collections.Specialized;
-using System.Globalization;
-using System.IO;
-using System.Linq;
-using System.Text;
-using System.Text.RegularExpressions;
-using System.Threading.Tasks;
 
 namespace cs_test
 {
