+++ cs_unittest/TestPooling.cs
@@ -1,12 +1,9 @@
-﻿using VW;
-using VW.Interfaces;
-using Microsoft.VisualStudio.TestTools.UnitTesting;
-using System;
+﻿using System;
 using System.Collections.Generic;
-using System.Linq;
-using System.Text;
 using System.Threading;
-using System.Threading.Tasks;
+using Microsoft.VisualStudio.TestTools.UnitTesting;
+using VW;
+using VW.Interfaces;
 
 namespace cs_unittest
 {
